+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GameMediaService.java
@@ -35,17 +35,18 @@ import de.mephisto.vpin.server.frontend.WheelIconDelete;
 import de.mephisto.vpin.server.highscores.HighscoreService;
 import de.mephisto.vpin.server.highscores.cards.CardService;
 import de.mephisto.vpin.server.listeners.EventOrigin;
-import de.mephisto.vpin.server.mame.MameRomAliasService;
-import de.mephisto.vpin.server.mame.MameService;
+import de.mephisto.vpin.server.vpinmame.VPinMameRomAliasService;
+import de.mephisto.vpin.server.vpinmame.VPinMameService;
 import de.mephisto.vpin.server.music.MusicService;
 import de.mephisto.vpin.server.pinvol.PinVolService;
 import de.mephisto.vpin.server.preferences.PreferencesService;
 import de.mephisto.vpin.server.puppack.PupPacksService;
 import de.mephisto.vpin.server.system.DefaultPictureService;
 import de.mephisto.vpin.server.vps.VpsService;
-import edu.umd.cs.findbugs.annotations.NonNull;
+import org.jspecify.annotations.NonNull;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.Strings;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -54,11 +55,12 @@ import org.springframework.stereotype.Service;
 
 import java.io.File;
 import java.io.IOException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
+import java.time.OffsetDateTime;
 
 @Service
 public class GameMediaService extends MediaService {
@@ -74,7 +76,7 @@ public class GameMediaService extends MediaService {
   private DMDDeviceIniService dmdDeviceIniService;
 
   @Autowired
-  private MameRomAliasService mameRomAliasService;
+  private VPinMameRomAliasService VPinMameRomAliasService;
 
   @Autowired
   private PinVolService pinVolService;
@@ -83,7 +85,7 @@ public class GameMediaService extends MediaService {
   private AssetRepository assetRepository;
 
   @Autowired
-  private MameService mameService;
+  private VPinMameService vPinMameService;
 
   @Autowired
   private GameService gameService;
@@ -226,13 +228,13 @@ public class GameMediaService extends MediaService {
           if (game.getVBSFile().exists()) {
             de.mephisto.vpin.restclient.util.FileUtils.renameToBaseName(game.getVBSFile(), name);
           }
-          LOG.info("Finished game file renaming from \"" + originalTableDetails.getGameFileName() + "\" to \"" + updatedTableDetails.getGameFileName() + "\"");
+          LOG.info("Finished game file renaming from \"{}\" to \"{}\"", originalTableDetails.getGameFileName(), updatedTableDetails.getGameFileName());
         }
         else {
           //revert to old value
           updatedTableDetails.setGameFileName(originalTableDetails.getGameFileName());
           frontendService.saveTableDetails(gameId, updatedTableDetails);
-          LOG.info("Renaming game file from \"" + originalTableDetails.getGameFileName() + "\" to \"" + updatedTableDetails.getGameFileName() + "\" failed, game file renaming failed.");
+          LOG.info("Renaming game file from \"{}\" to \"{}\" failed, game file renaming failed.", originalTableDetails.getGameFileName(), updatedTableDetails.getGameFileName());
         }
       }
     }
@@ -254,8 +256,8 @@ public class GameMediaService extends MediaService {
 
 
   public void runHighscoreRefreshCheck(Game game, TableDetails oldDetails, TableDetails newDetails) {
-    boolean romChanged = !StringUtils.equalsIgnoreCase(oldDetails.getRomName(), newDetails.getRomName());
-    boolean hsChanged = !StringUtils.equalsIgnoreCase(oldDetails.getHsFilename(), newDetails.getHsFilename());
+    boolean romChanged = !Strings.CI.equals(oldDetails.getRomName(), newDetails.getRomName());
+    boolean hsChanged = !Strings.CI.equals(oldDetails.getHsFilename(), newDetails.getHsFilename());
 
     if (romChanged || hsChanged) {
       LOG.info("Game highscore data fields have been changed, triggering score check.");
@@ -285,16 +287,17 @@ public class GameMediaService extends MediaService {
 
     //create backup first and delete existing table
     File existingVPXFile = new File(gameEmulator.getGamesDirectory(), tableDetails.getGameFileName());
-    long existingModifiationDate = existingVPXFile.lastModified();
+    long existingModificationDate = existingVPXFile.lastModified();
     if (existingVPXFile.exists()) {
       if (keepCopy) {
         File tableBackupsFolder = gameEmulator.getTableBackupsFolder();
         tableBackupsFolder.mkdirs();
-        String format = new SimpleDateFormat("yyyy-MM-dd-HH-mm-ss").format(new Date());
+        DateTimeFormatter df = DateTimeFormatter.ofPattern("yyyy-MM-dd-HH-mm-ss");
+        String format = df.format(OffsetDateTime.now());
         String suffix = FilenameUtils.getExtension(existingVPXFile.getName());
         File backup = new File(tableBackupsFolder, FilenameUtils.getBaseName(existingVPXFile.getName()) + "[" + format + "]." + suffix);
         org.apache.commons.io.FileUtils.copyFile(existingVPXFile, backup);
-        LOG.info("Created backup VPX file \"" + backup.getAbsolutePath() + "\"");
+        LOG.info("Created backup VPX file \"{}\"", backup.getAbsolutePath());
       }
 
       if (!existingVPXFile.delete()) {
@@ -310,7 +313,7 @@ public class GameMediaService extends MediaService {
     String baseName = FilenameUtils.getBaseName(tableDetails.getGameFileName());
     File existingVbsFile = new File(gameEmulator.getGamesDirectory(), baseName + ".vbs");
     if (existingVbsFile.exists() && !existingVbsFile.delete()) {
-      LOG.error("Failed to delete existing .vbs file \"" + existingVbsFile.getAbsolutePath() + "\"");
+      LOG.error("Failed to delete existing .vbs file \"{}\"", existingVbsFile.getAbsolutePath());
     }
 
     //Determine target name
@@ -330,9 +333,9 @@ public class GameMediaService extends MediaService {
 
     //keep modification date
     if (keepModificationDate) {
-      boolean b = target.setLastModified(existingModifiationDate);
+      boolean b = target.setLastModified(existingModificationDate);
       if (b) {
-        LOG.info("Reverted modification of VPX file \"{}\" to \"{}\"", temporaryVPXFile.getAbsolutePath(), new Date(existingModifiationDate));
+        LOG.info("Reverted modification of VPX file \"{}\" to \"{}\"", temporaryVPXFile.getAbsolutePath(), OffsetDateTime.ofInstant(Instant.ofEpochMilli(existingModificationDate), ZoneId.systemDefault()));
       }
       else {
         LOG.warn("Revetring modification of VPX file \"{}\" failed.", temporaryVPXFile.getAbsolutePath());
@@ -344,7 +347,7 @@ public class GameMediaService extends MediaService {
     baseName = FilenameUtils.getBaseName(target.getName());
     existingVbsFile = new File(target.getParentFile(), baseName + ".vbs");
     if (existingVbsFile.exists() && !existingVbsFile.delete()) {
-      LOG.error("Failed to delete .vbs file \"" + existingVbsFile.getAbsolutePath() + "\"");
+      LOG.error("Failed to delete .vbs file \"{}\"", existingVbsFile.getAbsolutePath());
     }
 
     //update frontend table database entry
@@ -364,7 +367,7 @@ public class GameMediaService extends MediaService {
     saveTableDetails(tableDetails, uploadDescriptor.getGameId(), !keepExistingFilename);
     frontendService.updateTableFileUpdated(uploadDescriptor.getGameId());
 
-    Game game = gameService.scanGame(uploadDescriptor.getGameId());
+    Game game = gameService.scanGame(uploadDescriptor.getGameId(), true);
     if (game != null) {
       gameService.resetUpdate(game.getId(), VpsDiffTypes.tableNewVPX);
       gameService.resetUpdate(game.getId(), VpsDiffTypes.tableNewVersionVPX);
@@ -380,7 +383,7 @@ public class GameMediaService extends MediaService {
 
       TableDataUtil.setMappedFieldValue(tableDetails, serverSettings.getMappingPatchVersion(), uploadDescriptor.getPatchVersion());
       frontendService.saveTableDetails(game.getId(), tableDetails);
-      LOG.info("Import of \"" + game.getGameDisplayName() + "\" successful.");
+      LOG.info("Import of \"{}\" successful.", game.getGameDisplayName());
     }
   }
 
@@ -407,13 +410,13 @@ public class GameMediaService extends MediaService {
     }
     targetVPXFile = FileUtils.uniqueFile(targetVPXFile);
 
-    LOG.info("Resolve target VPX: " + targetVPXFile.getAbsolutePath());
+    LOG.info("Resolve target VPX: {}", targetVPXFile.getAbsolutePath());
     org.apache.commons.io.FileUtils.copyFile(temporaryVPXFile, targetVPXFile);
-    LOG.info("Copied for import '" + temporaryVPXFile.getAbsolutePath() + "' to '" + targetVPXFile.getAbsolutePath() + "'");
+    LOG.info("Copied for import '{}' to '{}'", temporaryVPXFile.getAbsolutePath(), targetVPXFile.getAbsolutePath());
 
     int returningGameId = frontendService.importGame(targetVPXFile, true, -1, gameEmulator.getId());
     if (returningGameId >= 0) {
-      Game game = gameService.scanGame(returningGameId);
+      Game game = gameService.scanGame(returningGameId, true);
       if (game != null) {
         if (!uploadDescriptor.isBackupRestoreMode()) {
           if (uploadDescriptor.isAutoFill()) {
@@ -434,9 +437,8 @@ public class GameMediaService extends MediaService {
           TableDetails backedUpTableDetails = VpaArchiveUtil.readTableDetails(analysis.getFile());
           backedUpTableDetails.setGameName(newTableDetails.getGameName());
           backedUpTableDetails.setGameFileName(newTableDetails.getGameFileName());
-          backedUpTableDetails.setDateAdded(new Date());
+          backedUpTableDetails.setDateAdded(OffsetDateTime.now());
           backedUpTableDetails.setTourneyId(null);
-          backedUpTableDetails.setWebGameId(null);
           backedUpTableDetails.setLastPlayed(null);
           backedUpTableDetails.setEmulatorId(uploadDescriptor.getEmulatorId());
           frontendService.saveTableDetails(game.getId(), backedUpTableDetails);
@@ -471,7 +473,7 @@ public class GameMediaService extends MediaService {
 
 
         uploadDescriptor.setGameId(returningGameId);
-        LOG.info("Import of \"" + game.getGameDisplayName() + "\" successful.");
+        LOG.info("Import of \"{}\" successful.", game.getGameDisplayName());
       }
     }
   }
@@ -503,7 +505,7 @@ public class GameMediaService extends MediaService {
       target = new File(targetSubFolder, target.getName());
       fileName = targetSubFolder.getName() + "\\" + target.getName();
 
-      LOG.info("Clone of " + existingVPXFile.getName() + " is created into subfolder \"" + targetSubFolder.getAbsolutePath() + "\"");
+      LOG.info("Clone of {} is created into subfolder \"{}\"", existingVPXFile.getName(), targetSubFolder.getAbsolutePath());
     }
     else {
       target = FileUtils.uniqueFile(target);
@@ -512,12 +514,12 @@ public class GameMediaService extends MediaService {
 
     //copy file
     org.apache.commons.io.FileUtils.copyFile(temporaryVPXFile, target);
-    LOG.info("Copied temporary VPX file \"" + temporaryVPXFile.getAbsolutePath() + "\" to target \"" + target.getAbsolutePath() + "\"");
+    LOG.info("Copied temporary VPX file \"{}\" to target \"{}\"", temporaryVPXFile.getAbsolutePath(), target.getAbsolutePath());
 
 
     int returningGameId = frontendService.importGame(target, true, -1, gameEmulator.getId());
     if (returningGameId >= 0) {
-      Game importedGame = gameService.scanGame(returningGameId);
+      Game importedGame = gameService.scanGame(returningGameId, true);
       Game original = gameService.getGame(uploadDescriptor.getGameId());
 
       //update table details after new entry creation
@@ -530,10 +532,10 @@ public class GameMediaService extends MediaService {
 
       saveTableDetails(tableDetailsClone, returningGameId, false);
       frontendService.updateTableFileUpdated(returningGameId);
-      LOG.info("Created database clone entry with game name \"" + tableDetailsClone.getGameName() + "\"");
+      LOG.info("Created database clone entry with game name \"{}\"", tableDetailsClone.getGameName());
 
       //clone media
-      LOG.info("Cloning assets from game name \"" + original.getGameName() + "\" to \"" + importedGame.getGameName() + "\"");
+      LOG.info("Cloning assets from game name \"{}\" to \"{}\"", original.getGameName(), importedGame.getGameName());
       cloneGameMedia(original, importedGame);
 
       //clone additional files
@@ -553,7 +555,7 @@ public class GameMediaService extends MediaService {
       else {
         frontendService.vpsLink(importedGame.getId(), original.getExtTableId(), original.getExtTableVersionId());
       }
-      LOG.info("Cloning of \"" + importedGame.getGameDisplayName() + "\" successful.");
+      LOG.info("Cloning of \"{}\" successful.", importedGame.getGameDisplayName());
     }
 
     //update the game id to the new table
@@ -571,21 +573,21 @@ public class GameMediaService extends MediaService {
             String suffix = FilenameUtils.getExtension(mediaFile.getName());
             File cloneTarget = new File(frontendService.getMediaFolder(clone, originalScreenValue, suffix, true), clone.getGameName() + "." + suffix);
             if (mediaFile.getName().equals(cloneTarget.getName())) {
-              LOG.warn("Source name and target name of media asset " + mediaFile.getAbsolutePath() + " are identical, skipping cloning.");
+              LOG.warn("Source name and target name of media asset {} are identical, skipping cloning.", mediaFile.getAbsolutePath());
               return;
             }
 
             if (cloneTarget.exists() && !cloneTarget.delete()) {
-              LOG.error("Failed to clone media asset " + cloneTarget.getAbsolutePath() + ": deletion of existing asset failed.");
+              LOG.error("Failed to clone media asset {}: deletion of existing asset failed.", cloneTarget.getAbsolutePath());
               return;
             }
             org.apache.commons.io.FileUtils.copyFile(mediaFile, cloneTarget);
-            LOG.info("Cloned media asset: " + mediaFile.getAbsolutePath() + " to " + cloneTarget.getAbsolutePath());
+            LOG.info("Cloned media asset: {} to {}", mediaFile.getAbsolutePath(), cloneTarget.getAbsolutePath());
           }
         }
       }
       catch (IOException e) {
-        LOG.info("Failed to clone media asset: " + e.getMessage(), e);
+        LOG.info("Failed to clone media asset: {}", e.getMessage(), e);
       }
     }
   }
@@ -605,15 +607,15 @@ public class GameMediaService extends MediaService {
 
           if (de.mephisto.vpin.restclient.util.FileUtils.assetRename(gameMediaFile, oldBaseName, newBaseName)) {
             assetRenameCounter++;
-            LOG.info("[" + screen + "] Renamed media asset from \"" + gameMediaFile.getName() + "\" to name \"" + newBaseName + "\"");
+            LOG.info("[{}] Renamed media asset from \"{}\" to name \"{}\"", screen, gameMediaFile.getName(), newBaseName);
           }
           else {
-            LOG.warn("[" + screen + "] Renaming media asset from \"" + gameMediaFile.getName() + "\" to name \"" + newBaseName + "\" failed.");
+            LOG.warn("[{}] Renaming media asset from \"{}\" to name \"{}\" failed.", screen, gameMediaFile.getName(), newBaseName);
           }
         }
       }
     }
-    LOG.info("Finished asset renaming for \"" + oldBaseName + "\" to \"" + newBaseName + "\", renamed " + assetRenameCounter + " assets.");
+    LOG.info("Finished asset renaming for \"{}\" to \"{}\", renamed {} assets.", oldBaseName, newBaseName, assetRenameCounter);
   }
 
   public void installMediaPack(@NonNull UploadDescriptor uploadDescriptor, @NonNull UploaderAnalysis analysis) throws Exception {
@@ -640,14 +642,14 @@ public class GameMediaService extends MediaService {
         }
 
         if (PackageUtil.unpackTargetFile(tempFile, out, mediaFile)) {
-          LOG.info("Created \"" + out.getAbsolutePath() + "\" for screen \"" + screen.name() + "\" from archive file \"" + mediaFile + "\"");
+          LOG.info("Created \"{}\" for screen \"{}\" from archive file \"{}\"", out.getAbsolutePath(), screen.name(), mediaFile);
 
           if (game != null) {
             notifyGameScreenAssetsChanged(game.getId(), screen, out);
           }
         }
         else {
-          LOG.error("Failed to unpack " + out.getAbsolutePath() + " from " + tempFile.getAbsolutePath());
+          LOG.error("Failed to unpack {} from {}", out.getAbsolutePath(), tempFile.getAbsolutePath());
         }
 
         maxAssets--;
@@ -754,8 +756,8 @@ public class GameMediaService extends MediaService {
           }
         }
 
-        if (descriptor.isDeleteAlias()) {
-          if (!mameRomAliasService.deleteAlias(gameEmulator, game.getRomAlias())) {
+        if (descriptor.isDeleteAlias() && game.isVpxGame()) {
+          if (!VPinMameRomAliasService.deleteAlias(gameEmulator, game.getRomAlias())) {
             success = false;
           }
         }
@@ -788,19 +790,19 @@ public class GameMediaService extends MediaService {
         }
 
         if (descriptor.isDeleteDMDDeviceIni()) {
-          if (!mameService.deleteDMDDeviceIniEntry(game)) {
+          if (!vPinMameService.deleteDMDDeviceIniEntry(game)) {
             success = false;
           }
         }
 
         //cfg files belong to MAME
-        if (descriptor.isDeleteCfg()) {
-          if (!mameService.deleteCfg(game)) {
+        if (descriptor.isDeleteCfg() && game.isVpxGame()) {
+          if (!vPinMameService.deleteCfg(game)) {
             success = false;
           }
 
           if (!StringUtils.isEmpty(game.getRom())) {
-            if (!mameService.deleteOptions(game.getRom())) {
+            if (!vPinMameService.deleteOptions(game.getRom())) {
               success = false;
             }
           }
@@ -808,13 +810,13 @@ public class GameMediaService extends MediaService {
 
         if (descriptor.isDeleteRom()) {
           if (!StringUtils.isEmpty(game.getRom())) {
-            if (!mameService.deleteRom(game)) {
+            if (!vPinMameService.deleteRom(game)) {
               success = false;
             }
           }
         }
 
-        if (descriptor.isDeleteMusic()) {
+        if (descriptor.isDeleteMusic() && game.isVpxGame()) {
           if (!musicService.delete(game)) {
             success = false;
           }
@@ -841,13 +843,13 @@ public class GameMediaService extends MediaService {
             //only delete the assets, if there is no other game with the same "Game Name".
             List<Game> allOtherTables = this.frontendService.getGamesByEmulator(game.getEmulatorId())
                 .stream().filter(g -> g.getId() != game.getId())
-                .collect(Collectors.toList());
+                .toList();
             List<Game> duplicateGameNameTables = allOtherTables
                 .stream().filter(t -> t.getGameName().equalsIgnoreCase(game.getGameName()))
-                .collect(Collectors.toList());
+                .toList();
 
             if (duplicateGameNameTables.isEmpty()) {
-              LOG.info("Deleting screen assets for \"" + game.getGameDisplayName() + "\"");
+              LOG.info("Deleting screen assets for \"{}\"", game.getGameDisplayName());
               VPinScreen[] values = VPinScreen.values();
               for (VPinScreen originalScreenValue : values) {
                 List<FrontendMediaItem> frontendMediaItem = frontendService.getMediaItems(game, originalScreenValue);
@@ -881,7 +883,7 @@ public class GameMediaService extends MediaService {
           String[] list = gameFolder.list();
           if (list == null || list.length == 0) {
             if (gameFolder.delete()) {
-              LOG.info("Deleted table folder " + gameFolder.getAbsolutePath());
+              LOG.info("Deleted table folder {}", gameFolder.getAbsolutePath());
             }
           }
         }
